requirements.txt
# API and Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0

# Data Processing and Analysis
pandas==2.1.0
numpy==1.25.2
pandas-ta==0.3.14b0
matplotlib==3.7.3
seaborn==0.12.2

# KuCoin API
requests==2.31.0
aiohttp==3.8.5

# Scheduling and Async
APScheduler==3.10.4
asyncio==3.4.3

# Monitoring and Dashboard
dash==2.13.0
plotly==5.16.1
psutil==5.9.5

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
